requirements.txt
Django
djangorestframework
transformers
tensorflow